feat: :sparkles: #16 - Pass in existing values to functions
These changes allow the extension to remember what the user previously entered. Example, if they entered a description then went to the next step, then went back to edit the description, the user wouldn't have to remember what they previously entered to edit it. I have passed in the existing values to the prompts.